+++ CHANGELOG.md
@@ -11,6 +11,8 @@ For older change log history see the [historic changelog](HISTORIC_CHANGELOG.md)
 - ADManagedServiceAccount
   - New parameter TrustedForDelegation for Kerberos Delegation
     ([issue #717](https://github.com/dsccommunity/ActiveDirectoryDsc/issues/717)).
+  - New parameter TrustedForDelegation for Service Principal Name
+    ([issue #717](https://github.com/dsccommunity/ActiveDirectoryDsc/issues/717)).
 - ADDomainController
   - New parameter UseExistingAccount for attaching a server to an existing RODC account.
     ([issue #711](https://github.com/dsccommunity/ActiveDirectoryDsc/issues/711)).